--- src/components/Nodes.jsx
@@ -1,32 +1,0 @@
-export const onDragOver = useCallback((event) => {
-    event.preventDefault();
-    event.dataTransfer.dropEffect = 'move';
-}, []);
-
-export const onDrop = useCallback(
-    (event) => {
-        event.preventDefault();
-
-        const reactFlowBounds = reactFlowWrapper.current.getBoundingClientRect();
-        const type = event.dataTransfer.getData('application/reactflow');
-
-        // check if the dropped element is valid
-        if (typeof type === 'undefined' || !type) {
-            return;
-        }
-
-        const position = reactFlowInstance.project({
-            x: event.clientX - reactFlowBounds.left,
-            y: event.clientY - reactFlowBounds.top,
-        });
-        const newNode = {
-            id: getId(),
-            type,
-            position,
-            data: { label: `${type} node` },
-        };
-
-        setNodes((nds) => nds.concat(newNode));
-    },
-    [reactFlowInstance]
-);
